Remove redundant randomSelector
