+++ code/go/0chain.net/smartcontract/partitions/partitions.go
@@ -1,13 +1,19 @@
 package partitions
 
 import (
+	"bytes"
+	"errors"
 	"fmt"
 	"math/rand"
+	"reflect"
+	"strconv"
 
 	"0chain.net/core/common"
 	"0chain.net/core/datastore"
 	"0chain.net/core/encryption"
+	"github.com/0chain/common/core/logging"
 	"github.com/0chain/common/core/util"
+	"go.uber.org/zap"
 
 	"0chain.net/chaincore/chain/state"
 )
@@ -17,13 +23,20 @@ const (
 	errItemExistCode    = "item already exist"
 )
 
-// APIs
+const notFoundIndex = -1
+
+//msgp:ignore Partitions idIndex
+//go:generate msgp -io=false -tests=false -unexported=true -v
 
 type Partitions struct {
-	rs        *randomSelector
-	toRemove  []string
-	toAdd     []idIndex
-	locations map[string]int
+	Name          string       `json:"name"`
+	PartitionSize int          `json:"partition_size"`
+	NumPartitions int          `json:"num_partitions"`
+	Partitions    []*partition `json:"-" msg:"-"`
+
+	toRemove  []string       `json:"-" msg:"-"`
+	toAdd     []idIndex      `json:"-" msg:"-"`
+	locations map[string]int `json:"-" msg:"-"`
 }
 
 type idIndex struct {
@@ -40,23 +53,22 @@ type PartitionItem interface {
 // It's a common patten to call this to create partitions on
 // top-level partitions when project start
 func CreateIfNotExists(state state.StateContextI, name string, partitionSize int) (*Partitions, error) {
-	rs := randomSelector{}
-	err := state.GetTrieNode(name, &rs)
+	p := Partitions{}
+	err := state.GetTrieNode(name, &p)
 	switch err {
 	case nil:
-		return &Partitions{rs: &rs}, nil
+		return &p, nil
 	case util.ErrValueNotPresent:
-		rs, err := newRandomSelector(name, partitionSize)
+		p, err := newPartitions(name, partitionSize)
 		if err != nil {
 			return nil, err
 		}
 
-		pt := &Partitions{rs: rs}
-		if err := pt.rs.Save(state); err != nil {
+		if err := p.Save(state); err != nil {
 			return nil, err
 		}
 
-		return pt, nil
+		return p, nil
 	default:
 		return nil, err
 	}
@@ -64,12 +76,12 @@ func CreateIfNotExists(state state.StateContextI, name string, partitionSize int
 
 // GetPartitions returns partitions of given name
 func GetPartitions(state state.StateContextI, name string) (*Partitions, error) {
-	rs := randomSelector{}
-	if err := state.GetTrieNode(name, &rs); err != nil {
+	p := Partitions{}
+	if err := state.GetTrieNode(name, &p); err != nil {
 		return nil, err
 	}
 
-	return &Partitions{rs: &rs}, nil
+	return &p, nil
 }
 
 // ErrItemNotFound checks if error is common.Error and code is 'item not found'
@@ -92,58 +104,315 @@ func ErrItemExist(err error) bool {
 	return cErr.Code == errItemExistCode
 }
 
-// GetName returns the partitions name
-func (p *Partitions) GetName() string {
-	return p.rs.Name
+func newPartitions(name string, size int) (*Partitions, error) {
+	// TODO: limit the name length
+	return &Partitions{
+		Name:          name,
+		PartitionSize: size,
+	}, nil
 }
 
-// AddItem adds a partition item to parititons
-func (p *Partitions) AddItem(state state.StateContextI, item PartitionItem) error {
+func partitionKey(name string, index int) datastore.Key {
+	return name + encryption.Hash(":partition:"+strconv.Itoa(index))
+}
+
+func (p *Partitions) partitionKey(index int) datastore.Key {
+	return partitionKey(p.Name, index)
+}
+
+func (p *Partitions) Add(state state.StateContextI, item PartitionItem) (int, error) {
 	// duplicate item checking
 	_, ok, err := p.getItemPartIndex(state, item.GetID())
 	if err != nil {
-		return err
+		return 0, err
 	}
 
 	if ok {
-		return common.NewError(errItemExistCode, item.GetID())
+		return 0, common.NewError(errItemExistCode, item.GetID())
 	}
 
-	idx, err := p.rs.Add(state, item)
+	idx, err := p.add(state, item)
 	if err != nil {
-		return err
+		return 0, err
 	}
+
 	p.toAdd = append(p.toAdd, idIndex{
 		ID:  item.GetID(),
 		Idx: idx,
 	})
 
 	p.loadLocations(idx)
+	return idx, nil
+}
+
+func (p *Partitions) add(state state.StateContextI, item PartitionItem) (int, error) {
+	var (
+		part     *partition
+		err      error
+		partsNum = p.partitionsNum()
+	)
+
+	if partsNum > 0 {
+		part, err = p.getPartition(state, partsNum-1)
+		if err != nil {
+			return 0, err
+		}
+	}
+
+	if partsNum == 0 || part.length() >= p.PartitionSize {
+		part = p.addPartition()
+	}
+
+	if err := part.add(item); err != nil {
+		return 0, err
+	}
+
+	return p.partitionsNum() - 1, nil
+}
+
+func (p *Partitions) Get(state state.StateContextI, id string, v PartitionItem) error {
+	loc, ok, err := p.getItemPartIndex(state, id)
+	if err != nil {
+		return err
+	}
+
+	if !ok {
+		return common.NewError(errItemNotFoundCode, id)
+	}
+
+	if err := p.get(state, loc, id, v); err != nil {
+		return err
+	}
+
+	p.loadLocations(loc)
 	return nil
 }
 
-func (p *Partitions) loadLocations(idx int) {
-	if p.locations == nil {
-		p.locations = make(map[string]int)
-	}
-	if idx < 0 {
-		return
+func (p *Partitions) get(state state.StateContextI, partIndex int, id string, v PartitionItem) error {
+	pt, err := p.getPartition(state, partIndex)
+	if err != nil {
+		return err
 	}
 
-	part := p.rs.Partitions[idx]
-	for _, it := range part.Items {
-		kid := p.getLocKey(it.ID)
-		if _, ok := p.locations[kid]; ok {
-			return
-		}
-
-		p.locations[kid] = idx
+	item, _, ok := pt.find(id)
+	if !ok {
+		return errors.New("item not present")
 	}
+
+	_, err = v.UnmarshalMsg(item.Data)
+	return err
 }
 
-// Save saves the partitions data into state
+func (p *Partitions) UpdateItem(state state.StateContextI, it PartitionItem) error {
+	loc, ok, err := p.getItemPartIndex(state, it.GetID())
+	if err != nil {
+		return err
+	}
+
+	if !ok {
+		return common.NewError(errItemNotFoundCode, it.GetID())
+	}
+
+	if err := p.updateItem(state, loc, it); err != nil {
+		return err
+	}
+
+	p.loadLocations(loc)
+	return nil
+}
+
+func (p *Partitions) updateItem(
+	state state.StateContextI,
+	partIndex int,
+	it PartitionItem,
+) error {
+	part, err := p.getPartition(state, partIndex)
+	if err != nil {
+		return err
+	}
+
+	return part.update(it)
+}
+
+func (p *Partitions) Update(state state.StateContextI, key string, f func(data []byte) ([]byte, error)) error {
+	l, ok, err := p.getItemPartIndex(state, key)
+	if err != nil {
+		return err
+	}
+
+	if !ok {
+		return common.NewError(errItemNotFoundCode, key)
+	}
+
+	part, err := p.getPartition(state, l)
+	if err != nil {
+		return err
+	}
+
+	v, idx, ok := part.find(key)
+	if !ok {
+		return common.NewError(errItemNotFoundCode, key)
+	}
+
+	nData, err := f(v.Data)
+	if err != nil {
+		return err
+	}
+	v.Data = nData
+	part.Items[idx] = v
+	part.Changed = true
+
+	p.loadLocations(l)
+	return nil
+}
+
+func (p *Partitions) Remove(state state.StateContextI, id string) error {
+	loc, ok, err := p.getItemPartIndex(state, id)
+	if err != nil {
+		return err
+	}
+
+	if !ok {
+		return common.NewError(errItemNotFoundCode, id)
+	}
+
+	if err := p.removeItem(state, id, loc); err != nil {
+		return err
+	}
+
+	p.toRemove = append(p.toRemove, id)
+
+	p.loadLocations(loc)
+	delete(p.locations, p.getLocKey(id))
+	return nil
+
+}
+
+func (p *Partitions) removeItem(
+	state state.StateContextI,
+	id string,
+	index int,
+) error {
+	part, err := p.getPartition(state, index)
+	if err != nil {
+		return err
+	}
+
+	err = part.remove(id)
+	if err != nil {
+		return err
+	}
+
+	lastPart, err := p.getPartition(state, p.partitionsNum()-1)
+	if err != nil {
+		return err
+	}
+
+	if index == p.partitionsNum()-1 {
+		if lastPart.length() == 0 {
+			if err := p.deleteTail(state); err != nil {
+				return err
+			}
+		}
+		return nil
+	}
+
+	replace := lastPart.cutTail()
+	if replace == nil {
+		logging.Logger.Error("empty last partition - should not happen!!",
+			zap.Int("part index", p.NumPartitions-1),
+			zap.Int("part num", p.NumPartitions),
+			zap.Int("parts len", p.NumPartitions))
+
+		return fmt.Errorf("empty last partitions, currpt data")
+	}
+	if err := part.addRaw(*replace); err != nil {
+		return err
+	}
+
+	if lastPart.length() == 0 {
+		if err := p.deleteTail(state); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (p *Partitions) GetRandomItems(state state.StateContextI, r *rand.Rand, vs interface{}) error {
+	if p.partitionsNum() == 0 {
+		return errors.New("empty list, no items to return")
+	}
+	index := r.Intn(p.partitionsNum())
+
+	part, err := p.getPartition(state, index)
+	if err != nil {
+		return err
+	}
+
+	its, err := part.itemRange(0, part.length())
+	if err != nil {
+		return err
+	}
+
+	rtv := make([]item, 0, p.PartitionSize)
+	rtv = append(rtv, its...)
+
+	if index == p.partitionsNum()-1 && len(rtv) < p.PartitionSize && p.partitionsNum() > 1 {
+		secondLast, err := p.getPartition(state, index-1)
+		if err != nil {
+			return err
+		}
+		want := p.PartitionSize - len(rtv)
+		if secondLast.length() < want {
+			return fmt.Errorf("second last part too small %d instead of %d",
+				secondLast.length(), p.partitionsNum())
+		}
+		its, err := secondLast.itemRange(secondLast.length()-want, secondLast.length())
+		if err != nil {
+			return err
+		}
+
+		rtv = append(rtv, its...)
+	}
+
+	return setPartitionItems(rtv, vs)
+}
+
+func (p *Partitions) Size(state state.StateContextI) (int, error) {
+	if p.partitionsNum() == 0 {
+		return 0, nil
+	}
+
+	lastPart, err := p.getPartition(state, p.partitionsNum()-1)
+	if err != nil {
+		return 0, err
+	}
+
+	return (p.partitionsNum()-1)*p.PartitionSize + lastPart.length(), nil
+}
+
+func (p *Partitions) Exist(state state.StateContextI, id string) (bool, error) {
+	_, ok, err := p.getItemPartIndex(state, id)
+	if err != nil {
+		return false, err
+	}
+
+	return ok, nil
+}
+
 func (p *Partitions) Save(state state.StateContextI) error {
-	if err := p.rs.Save(state); err != nil {
+	for _, partition := range p.Partitions {
+		if partition != nil && partition.changed() {
+			err := partition.save(state)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	_, err := state.InsertTrieNode(p.Name, p)
+	if err != nil {
 		return err
 	}
 
@@ -165,184 +434,142 @@ func (p *Partitions) Save(state state.StateContextI) error {
 	return nil
 }
 
-// GetItem returns partition item of given partition index and id
-func (p *Partitions) GetItem(state state.StateContextI, id string, v PartitionItem) error {
-	loc, ok, err := p.getItemPartIndex(state, id)
-	if err != nil {
-		return err
-	}
+func (p *Partitions) foreach(state state.StateContextI, f func(string, []byte, int) ([]byte, bool, error)) error {
+	for i := 0; i < p.partitionsNum(); i++ {
+		part, err := p.getPartition(state, i)
+		if err != nil {
+			return fmt.Errorf("could not get partition: name:%s, index: %d", p.Name, i)
+		}
 
-	if !ok {
-		return common.NewError(errItemNotFoundCode, id)
-	}
+		for i, v := range part.Items {
+			ret, bk, err := f(v.ID, v.Data, i)
+			if err != nil {
+				return err
+			}
+			if !bytes.Equal(ret, v.Data) {
+				v.Data = ret
+				part.Items[i] = v
+				part.Changed = true
+			}
 
-	if err := p.rs.GetItem(state, loc, id, v); err != nil {
-		return err
+			if bk {
+				return nil
+			}
+		}
 	}
-
-	p.loadLocations(loc)
 
 	return nil
 }
 
-// UpdateItem updates item on given partition index
-func (p *Partitions) UpdateItem(state state.StateContextI, item PartitionItem) error {
-	loc, ok, err := p.getItemPartIndex(state, item.GetID())
-	if err != nil {
-		return err
+func setPartitionItems(rtv []item, vs interface{}) error {
+	// slice type
+	vst := reflect.TypeOf(vs)
+	if vst.Kind() != reflect.Ptr {
+		return errors.New("invalid return value type, it must be a pointer of slice")
 	}
 
-	if !ok {
-		return common.NewError(errItemNotFoundCode, item.GetID())
+	// element type - slice
+	vts := vst.Elem()
+	if vts.Kind() != reflect.Slice {
+		return errors.New("invalid return value type, it must be a pointer of slice")
 	}
 
-	if err := p.rs.UpdateItem(state, loc, item); err != nil {
-		return err
-	}
+	// item type
+	vt := vts.Elem()
 
-	p.loadLocations(loc)
-	return nil
-}
+	// create a new item slice
+	rv := reflect.MakeSlice(vts, len(rtv), len(rtv))
 
-// Size returns the total item number in partitions
-func (p *Partitions) Size(state state.StateContextI) (int, error) {
-	return p.rs.Size(state)
-}
+	for i, v := range rtv {
+		// create new item instance and assert PartitionItem interface
+		pi, ok := reflect.New(vt).Interface().(PartitionItem)
+		if !ok {
+			return errors.New("invalid value type, the item does not meet PartitionItem interface")
+		}
 
-// Num returns the number of partitions
-func (p *Partitions) Num() int {
-	return p.rs.partitionsNum()
-}
-
-// RemoveItem removes the partition item from given partIndex in Partitions
-func (p *Partitions) RemoveItem(state state.StateContextI, id string) error {
-	loc, ok, err := p.getItemPartIndex(state, id)
-	if err != nil {
-		return err
-	}
-
-	if !ok {
-		return common.NewError(errItemNotFoundCode, id)
-	}
-
-	if err := p.rs.RemoveItem(state, id, loc); err != nil {
-		return err
-	}
-
-	p.toRemove = append(p.toRemove, id)
-
-	p.loadLocations(loc)
-	delete(p.locations, p.getLocKey(id))
-	return nil
-}
-
-// RemoveItems removes items by keys from the same partition of partIndex
-func (p *Partitions) RemoveItems(state state.StateContextI, partIndex int, keys []string) error {
-	for _, k := range keys {
-		if err := p.rs.RemoveItem(state, k, partIndex); err != nil {
+		// decode data
+		if _, err := pi.UnmarshalMsg(v.Data); err != nil {
 			return err
 		}
-		p.toRemove = append(p.toRemove, k)
-		delete(p.locations, p.getLocKey(k))
+
+		// set to slice
+		rv.Index(i).Set(reflect.ValueOf(pi).Elem())
 	}
 
+	// set slice back to v param
+	reflect.ValueOf(vs).Elem().Set(rv)
 	return nil
 }
 
-// GetRandomItems returns items of partition size number from random partition,
-// if the last partition is not full, it will try to get and fill it with its partition
-// of index - 1.
-func (p *Partitions) GetRandomItems(state state.StateContextI, r *rand.Rand, v interface{}) error {
-	return p.rs.GetRandomItems(state, r, v)
+func (p *Partitions) addPartition() *partition {
+	newPartition := &partition{
+		Key: p.partitionKey(p.partitionsNum()),
+	}
+
+	p.Partitions = append(p.Partitions, newPartition)
+	p.NumPartitions++
+	return newPartition
 }
 
-type RandMatchFunc func(key string, data []byte) bool
-
-func (p *Partitions) UpdateRandomItems(state state.StateContextI, r *rand.Rand, randN int, f func(string, []byte) ([]byte, error)) error {
-	return p.rs.UpdateRandomItems(state, r, randN, f)
-}
-
-// Foreach loads all partitions and iterate through one by one
-// break whenever the callback function returns error
-// for the callback function, if it returns bytes different from the input data, then the
-// changes of items will be saved to partitions.
-func (p *Partitions) Foreach(state state.StateContextI, f func(key string, data []byte, partIndex int) ([]byte, bool, error)) error {
-	return p.rs.foreach(state, f)
-}
-
-func (p *Partitions) Exist(state state.StateContextI, id string) (bool, error) {
-	_, ok, err := p.getItemPartIndex(state, id)
+func (p *Partitions) deleteTail(balances state.StateContextI) error {
+	_, err := balances.DeleteTrieNode(p.partitionKey(p.partitionsNum() - 1))
 	if err != nil {
-		return false, err
-	}
-
-	return ok, nil
-}
-
-type ChangePartitionCallback = func(string, []byte, int, int, state.StateContextI) error
-
-func (p *Partitions) getLocKey(id string) datastore.Key {
-	return encryption.Hash(fmt.Sprintf("%s:%s", p.rs.Name, id))
-}
-
-func (p *Partitions) getItemPartIndex(state state.StateContextI, id string) (int, bool, error) {
-	var pl location
-
-	kid := p.getLocKey(id)
-	loc, ok := p.locations[kid]
-	if ok {
-		return loc, true, nil
-	}
-
-	if err := state.GetTrieNode(kid, &pl); err != nil {
-		if err == util.ErrValueNotPresent {
-			return -1, false, nil
+		if err != util.ErrValueNotPresent {
+			return err
 		}
-
-		return -1, false, err
 	}
-
-	return pl.Location, true, nil
-}
-
-func (p *Partitions) saveItemLoc(state state.StateContextI, id string, partIndex int) error {
-	_, err := state.InsertTrieNode(p.getLocKey(id), &location{Location: partIndex})
-	return err
-}
-
-func (p *Partitions) removeItemLoc(state state.StateContextI, id string) error {
-	_, err := state.DeleteTrieNode(p.getLocKey(id))
-	return err
-}
-
-func (p *Partitions) Update(state state.StateContextI, key string, f func(data []byte) ([]byte, error)) error {
-	l, ok, err := p.getItemPartIndex(state, key)
-	if err != nil {
-		return err
-	}
-
-	if !ok {
-		return common.NewError(errItemNotFoundCode, key)
-	}
-
-	part, err := p.rs.getPartition(state, l)
-	if err != nil {
-		return err
-	}
-
-	v, idx, ok := part.find(key)
-	if !ok {
-		return common.NewError(errItemNotFoundCode, key)
-	}
-
-	nData, err := f(v.Data)
-	if err != nil {
-		return err
-	}
-	v.Data = nData
-	part.Items[idx] = v
-	part.Changed = true
-
-	p.loadLocations(l)
+	p.Partitions = p.Partitions[:p.partitionsNum()-1]
+	p.NumPartitions--
 	return nil
 }
+
+func (p *Partitions) getPartition(state state.StateContextI, i int) (*partition, error) {
+	if i >= p.partitionsNum() {
+		return nil, fmt.Errorf("partition id %v greater than number of partitions %v", i, p.partitionsNum())
+	}
+	if p.Partitions[i] != nil {
+		return p.Partitions[i], nil
+	}
+
+	part := &partition{}
+	err := part.load(state, p.partitionKey(i))
+	if err != nil {
+		return nil, err
+	}
+	p.Partitions[i] = part
+	return part, nil
+}
+
+func (p *Partitions) partitionsNum() int {
+	// assert the partitions number match
+	if p.NumPartitions != len(p.Partitions) {
+		logging.Logger.DPanic(fmt.Sprintf("number of partitions mismatch, numPartitions: %d, len(partitions): %d",
+			p.NumPartitions, len(p.Partitions)))
+	}
+	return p.NumPartitions
+}
+
+func (p *Partitions) MarshalMsg(o []byte) ([]byte, error) {
+	d := partitionsDecode(*p)
+	return d.MarshalMsg(o)
+}
+
+func (p *Partitions) UnmarshalMsg(b []byte) ([]byte, error) {
+	d := &partitionsDecode{}
+	o, err := d.UnmarshalMsg(b)
+	if err != nil {
+		return nil, err
+	}
+
+	*p = Partitions(*d)
+
+	p.Partitions = make([]*partition, d.NumPartitions)
+	return o, nil
+}
+
+func (p *Partitions) Msgsize() int {
+	d := partitionsDecode(*p)
+	return d.Msgsize()
+}
+
+type partitionsDecode Partitions
