+++ code/go/0chain.net/smartcontract/partitions/partitions_test.go
@@ -1,26 +1,95 @@
 package partitions
 
 import (
+	"encoding/json"
 	"fmt"
-	"math/rand"
 	"testing"
-	"time"
 
+	"0chain.net/chaincore/chain/state/mocks"
+	"github.com/0chain/common/core/util"
 	"github.com/stretchr/testify/require"
 )
 
+type StringItem string
+
+func (si StringItem) GetID() string {
+	return string(si)
+}
+
+func (si *StringItem) MarshalMsg(o []byte) ([]byte, error) {
+	return []byte(*si), nil
+}
+
+func (si *StringItem) UnmarshalMsg(b []byte) ([]byte, error) {
+	*si = StringItem(b)
+	return nil, nil
+}
+
+func (si *StringItem) Msgsize() int {
+	return len(*si)
+}
+
+func ItemFromString(name string) PartitionItem {
+	v := StringItem(name)
+	return &v
+}
+
+type mockStateContextI struct {
+	*mocks.StateContextI
+	data map[string][]byte
+}
+
+func (m *mockStateContextI) GetTrieNode(key string, v util.MPTSerializable) error {
+	d, ok := m.data[key]
+	if !ok {
+		return util.ErrValueNotPresent
+	}
+
+	_, err := v.UnmarshalMsg(d)
+	return err
+}
+
+func (m *mockStateContextI) InsertTrieNode(key string, node util.MPTSerializable) (string, error) {
+	d, err := node.MarshalMsg(nil)
+	if err != nil {
+		return "", err
+	}
+	m.data[key] = d
+	return "", nil
+}
+
+type testItem struct {
+	ID string
+	V  string
+}
+
+func (ti *testItem) GetID() string {
+	return ti.ID
+}
+
+func (ti *testItem) MarshalMsg(o []byte) ([]byte, error) {
+	return json.Marshal(ti)
+}
+
+func (ti *testItem) UnmarshalMsg(b []byte) ([]byte, error) {
+	return nil, json.Unmarshal(b, ti)
+}
+
+func (ti *testItem) Msgsize() int {
+	d, _ := ti.MarshalMsg(nil)
+	return len(d)
+}
+
 func TestPartitionsSave(t *testing.T) {
 	balances := &mockStateContextI{data: make(map[string][]byte)}
-	rs, err := newRandomSelector("test_rs", 10, nil)
+	parts, err := newPartitions("test_rs", 10)
 	require.NoError(t, err)
-
-	parts := Partitions{rs: rs}
 
 	for i := 0; i < 20; i++ {
 		k := fmt.Sprintf("k%d", i)
 		v := fmt.Sprintf("v%d", i)
 		it := testItem{ID: k, V: v}
-		err = parts.AddItem(balances, &it)
+		_, err = parts.Add(balances, &it)
 		require.NoError(t, err)
 	}
 
@@ -31,91 +100,39 @@ func TestPartitionsSave(t *testing.T) {
 	require.NoError(t, err)
 
 	var v testItem
-	err = p1.GetItem(balances, "k15", &v)
+	err = p1.Get(balances, "k15", &v)
 	require.NoError(t, err)
 	require.Equal(t, "v15", v.V)
 
 	p2, err := GetPartitions(balances, "test_rs")
 	require.NoError(t, err)
-	require.Equal(t, 2, p2.rs.NumPartitions)
+	require.Equal(t, 2, p2.NumPartitions)
 
-	// update item
+	// updateItem item
 	err = p1.UpdateItem(balances, &testItem{"k10", "vv10"})
 	require.NoError(t, err)
 	require.NoError(t, p1.Save(balances))
 
 	p3, err := GetPartitions(balances, "test_rs")
 	require.NoError(t, err)
-	require.Equal(t, 2, p3.rs.NumPartitions)
+	require.Equal(t, 2, p3.NumPartitions)
 
 	var vv testItem
-	err = p3.GetItem(balances, "k10", &vv)
+	err = p3.Get(balances, "k10", &vv)
 	require.NoError(t, err)
 	require.Equal(t, "vv10", vv.V)
 }
 
-func Test_randomSelector_UpdateRandomItems(t *testing.T) {
-	balances := &mockStateContextI{data: make(map[string][]byte)}
-	rs, err := CreateIfNotExists(balances, "test_rs", 10)
-
-	items := make([]testItem, 0, 20)
-	for i := 0; i < 15; i++ {
-		k := fmt.Sprintf("k%d", i)
-		v := fmt.Sprintf("v%d", i)
-		it := testItem{ID: k, V: v}
-		items = append(items, it)
-		err = rs.AddItem(balances, &it)
-		require.NoError(t, err)
-	}
-
-	err = rs.Save(balances)
-	require.NoError(t, err)
-
-	rs2, err := GetPartitions(balances, "test_rs")
-	require.NoError(t, err)
-
-	r := rand.New(rand.NewSource(time.Now().Unix()))
-	var ids []string
-	err = rs2.UpdateRandomItems(balances, r, 8, func(key string, data []byte) ([]byte, error) {
-		ti := testItem{}
-		_, err := ti.UnmarshalMsg(data)
-		require.NoError(t, err)
-
-		fmt.Println(ti)
-
-		ti.V = ti.V + ":new added"
-		ids = append(ids, ti.ID)
-		return ti.MarshalMsg(nil)
-	})
-
-	require.NoError(t, err)
-
-	err = rs2.Save(balances)
-	require.NoError(t, err)
-
-	rs3, err := GetPartitions(balances, "test_rs")
-	require.NoError(t, err)
-
-	for _, id := range ids {
-		var item testItem
-		err = rs3.GetItem(balances, id, &item)
-		require.NoError(t, err)
-		fmt.Println(item)
-	}
-}
-
 func TestPartitionsForeach(t *testing.T) {
 	balances := &mockStateContextI{data: make(map[string][]byte)}
-	rs, err := newRandomSelector("test_rs", 10, nil)
+	parts, err := newPartitions("test_rs", 10)
 	require.NoError(t, err)
-
-	parts := Partitions{rs: rs}
 
 	for i := 0; i < 20; i++ {
 		k := fmt.Sprintf("k%d", i)
 		v := fmt.Sprintf("v%d", i)
 		it := testItem{ID: k, V: v}
-		_, err = parts.AddItem(balances, &it)
+		_, err = parts.Add(balances, &it)
 		require.NoError(t, err)
 	}
 
@@ -125,7 +142,7 @@ func TestPartitionsForeach(t *testing.T) {
 	p1, err := GetPartitions(balances, "test_rs")
 	require.NoError(t, err)
 
-	err = p1.Foreach(balances, func(key string, data []byte, _ int) ([]byte, bool, error) {
+	err = p1.foreach(balances, func(key string, data []byte, _ int) ([]byte, bool, error) {
 		if key == "k1" {
 			n := testItem{}
 			_, err := n.UnmarshalMsg(data)
@@ -149,30 +166,7 @@ func TestPartitionsForeach(t *testing.T) {
 	p2, err := GetPartitions(balances, "test_rs")
 	require.NoError(t, err)
 	vv := testItem{}
-	_, err = p2.GetItem(balances, "k1", &vv)
+	err = p2.Get(balances, "k1", &vv)
 	require.NoError(t, err)
 	require.Equal(t, "new item", vv.V)
-
-	//var v testItem
-	//_, err = p1.GetItem(balances, "k15", &v)
-	//require.NoError(t, err)
-	//require.Equal(t, "v15", v.V)
-	//
-	//p2, err := GetPartitions(balances, "test_rs")
-	//require.NoError(t, err)
-	//require.Equal(t, 2, p2.rs.NumPartitions)
-	//
-	//// update item
-	//err = p1.UpdateItem(balances, &testItem{"k10", "vv10"})
-	//require.NoError(t, err)
-	//require.NoError(t, p1.Save(balances))
-	//
-	//p3, err := GetPartitions(balances, "test_rs")
-	//require.NoError(t, err)
-	//require.Equal(t, 2, p3.rs.NumPartitions)
-	//
-	//var vv testItem
-	//_, err = p3.GetItem(balances, "k10", &vv)
-	//require.NoError(t, err)
-	//require.Equal(t, "vv10", vv.V)
 }
